[refactor] userexh 위치 이동
