+++ src/main/java/klieme/artdiary/calendar/service/CalendarService.java
@@ -16,9 +16,7 @@ import klieme.artdiary.calendar.info.ScheduleInfo;
 import klieme.artdiary.common.ImageTransfer;
 import klieme.artdiary.common.UserIdFilter;
 import klieme.artdiary.exhibitions.data_access.entity.ExhEntity;
-import klieme.artdiary.exhibitions.data_access.entity.UserExhEntity;
 import klieme.artdiary.exhibitions.data_access.repository.ExhRepository;
-import klieme.artdiary.exhibitions.data_access.repository.UserExhRepository;
 import klieme.artdiary.gatherings.data_access.entity.GatheringEntity;
 import klieme.artdiary.gatherings.data_access.entity.GatheringExhEntity;
 import klieme.artdiary.gatherings.data_access.entity.GatheringMateEntity;
@@ -26,6 +24,8 @@ import klieme.artdiary.gatherings.data_access.entity.GatheringMateId;
 import klieme.artdiary.gatherings.data_access.repository.GatheringExhRepository;
 import klieme.artdiary.gatherings.data_access.repository.GatheringMateRepository;
 import klieme.artdiary.gatherings.data_access.repository.GatheringRepository;
+import klieme.artdiary.myexhs.data_access.entity.UserExhEntity;
+import klieme.artdiary.myexhs.data_access.repository.UserExhRepository;
 
 @Service
 public class CalendarService implements CalendarReadUseCase {
