+++ src/main/java/klieme/artdiary/exhibitions/service/ExhService.java
@@ -15,9 +15,7 @@ import klieme.artdiary.common.ImageTransfer;
 import klieme.artdiary.common.MessageType;
 import klieme.artdiary.common.UserIdFilter;
 import klieme.artdiary.exhibitions.data_access.entity.ExhEntity;
-import klieme.artdiary.exhibitions.data_access.entity.UserExhEntity;
 import klieme.artdiary.exhibitions.data_access.repository.ExhRepository;
-import klieme.artdiary.exhibitions.data_access.repository.UserExhRepository;
 import klieme.artdiary.exhibitions.enums.ExhField;
 import klieme.artdiary.exhibitions.enums.ExhPrice;
 import klieme.artdiary.exhibitions.enums.ExhState;
@@ -34,6 +32,8 @@ import klieme.artdiary.gatherings.data_access.repository.GatheringMateRepository
 import klieme.artdiary.gatherings.data_access.repository.GatheringRepository;
 import klieme.artdiary.mydiarys.data_access.entity.MydiaryEntity;
 import klieme.artdiary.mydiarys.data_access.repository.MydiaryRepository;
+import klieme.artdiary.myexhs.data_access.entity.UserExhEntity;
+import klieme.artdiary.myexhs.data_access.repository.UserExhRepository;
 import klieme.artdiary.users.data_access.entity.UserEntity;
 import klieme.artdiary.users.data_access.repository.UserRepository;
