+++ src/main/java/klieme/artdiary/myexhs/data_access/entity/UserExhEntity.java
@@ -1,4 +1,4 @@
-package klieme.artdiary.exhibitions.data_access.entity;
+package klieme.artdiary.myexhs.data_access.entity;
 
 import java.time.LocalDate;
