+++ src/main/java/klieme/artdiary/myexhs/data_access/repository/UserExhRepository.java
@@ -1,4 +1,4 @@
-package klieme.artdiary.exhibitions.data_access.repository;
+package klieme.artdiary.myexhs.data_access.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import klieme.artdiary.exhibitions.data_access.entity.UserExhEntity;
+import klieme.artdiary.myexhs.data_access.entity.UserExhEntity;
 
 @Repository
 public interface UserExhRepository extends JpaRepository<UserExhEntity, Long> {
