+++ src/main/java/klieme/artdiary/users/service/UserService.java
@@ -14,10 +14,10 @@ import klieme.artdiary.common.ImageTransfer;
 import klieme.artdiary.common.ImageType;
 import klieme.artdiary.common.MessageType;
 import klieme.artdiary.common.UserIdFilter;
-import klieme.artdiary.exhibitions.data_access.entity.UserExhEntity;
-import klieme.artdiary.exhibitions.data_access.repository.UserExhRepository;
 import klieme.artdiary.gatherings.data_access.entity.GatheringDiaryEntity;
 import klieme.artdiary.gatherings.data_access.repository.GatheringDiaryRepository;
+import klieme.artdiary.myexhs.data_access.entity.UserExhEntity;
+import klieme.artdiary.myexhs.data_access.repository.UserExhRepository;
 import klieme.artdiary.users.data_access.entity.ReasonEntity;
 import klieme.artdiary.users.data_access.entity.UserEntity;
 import klieme.artdiary.users.data_access.repository.ReasonRepository;
